// src/types/dashboard/store.types.ts
export interface AllStoresTypes {
  already_have_domain: boolean;
  created_at: string;
  custom_domain: string | null;
  domain_name: string;
  id: string;
  industry_type: string;
  product_type: string;
  status: string;
  store_abbreviation: string;
  store_code: string;
  store_contact: string | null;
  store_country: string | null;
  store_description: string | null;
  store_location: string | null;
  store_logo: string | undefined;
  store_name: string;
  updated_at: string;
  user_id: number;
}